[Async CC] Auth'd AFP ptr of dynamically thin fn.
In irgen::getAsyncFunctionAndSize, in the case where the passed-in
function has a thick representation but is dynamically thin, the pointer
that is passed in is not actually a pointer to a function but instead a
pointer to an AsyncFunctionPointer struct.  To obtain the function
pointer and size, fields must be loaded from this struct.  However, in
order to load these fields, that pointer to the AsyncFunctionPointer
struct must itself first be authenticated.  Here, that authentication is
performed.