[Async CC] Resigning AFP results in AFP.
Previously, when a FunctionPointer whose kind was AsyncFunctionPointer
was resigned, the underlying function pointer was loaded from the
AsyncFunctionPointer struct and that value was signed.  That was
incorrect: a resigned pointer to an AsyncFunctionPointer struct should
again be a pointer to an AsyncFunctionPointer struct.  Here that error
is corrected and the result of resigning is again a pointer to an
AsyncFunctionPointer struct.