[Async CC] Resigned ptr passed to swift_task_create[_f].
When emitting Builtin.createAsyncTask[Future], the function passed to
the builtin is passed along to swift_task_create[_f].  The latter is
expecting a AsyncFunctionType<void()>, which is an alias for

  void (AsyncTask *, ExecutorRef, AsyncContext *)

Previously, no ptrauth logic was emitted when emitting the builtins, so
the function pointer was not signed in the way expected by
swift_task_create[_f].  The result was a ptrauth failure on arm64e.
Here, that problem is fixed by resigning the function pointer in the way
the runtime expects.