[PtrAuth] Pass discriminator to swift_suspend_dispatch.
Previously, swift_suspend_dispatch was passed a pointer to a function
and createAsyncDispatchFn was passed a FunctionPointer.  The latter
constructed a new FunctionPointer using the passed-in function pointer
as the value, because the value inside the FunctionPointer was a value
in a different function.

That worked fine on platforms without pointer authentication.  On
arm64e, however, calling a function can use two values from a
FunctionPointer: the pointer to the function and the discriminator.  The
result was that on arm64e, swift_suspend_dispatch failed verification
because the discriminator value that was used in the call made by
swift_suspend_dispatch did not originate in that function.

Here, that problem is resolved by passing the discriminator to
swift_suspend_dispatch.  Now, createAsyncDispatchFn creates a
FunctionPointer using not just the passed-in function pointer but also
the passed-in discriminator.  The result is that swift_suspend_dispatch
no longer fails verification.