[eager-spec] Teach how to handle undef arguments in the return bb.
Due to the pass order, this can run in between SemanticARCOpts and
SimplifyCFG. SemanticARCOpts will leave behind certain SILArguments that only
have undef arguments due to it having eliminated ownership instructions. The
idea was to leave that for SimplifyCFG to fixup so the pass would not invalidate
the CFG.

As a result, I exposed a bug in the eager specializer where if it pattern
matched:

```
bbFunctionExit(%arg : $Builtin.Int32):
  %9999 = tuple()
  return %9999
```

It would ignore the argument and branch to bbFunctionExit without any branch
arguments breaking CFG invariants. Now the pass bails if the merge block has any
SILArguments that do not have /all/ undef incoming values.