[IRGen] Use specialized mangled name for class constants.
Previously, various class constants for prespecialized generic metadata
such as _INSTANCE_METHODS_ and _DATA_ were mangled using the objc type
name of the class.  As a result, names differed from one another by a
number appended at the end of the string.  That resulted in symbols
which were difficult to work with.  Here, that issue is resolved by
appending the mangled name of the type metadata to the prefix.

rdar://problem/72053639