NFC: Refactor inout error msg emission
Global actors need to work with non-lookup non-member values. As such,
we can't rely on the argument of the call being a LookupExpr. The error
message emission logic only relies on the declaration and the subExpr
argument. This refactor will allow us to avoid having another switch
statement for the global-actor case that more-or-less does the same
thing.