stubs: provide an out-of-line definition of `sqrtf` on Win x86
`ucrt` does not provide a definition for `sqrtf` in the runtime
component.  It is expected that the symbol will be expanded by the
inline definition to lower to `sqrt` with the argument widened and the
result narrowed.  However, this requires the Swift shims to depend on
`ucrt` which is not permissible as it exists to break the circular
dependency in the modules.  This provides an out-of-line alias for the
symbol to allow the Swift runtime to resolve the symbol.