[SILGen] Fix crash when compiling attributed Clang function types.
Before this change, @convention(c|block) function types with non-trivial
Clang types would not be compiled correctly, because the AbstractionPattern
would sometimes be of Kind::Type instead of Kind::ClangType. We normalize
the AbstractionPattern right before lowering parameters and results to make
sure that any attributes on the stored Clang type are respected.