[Async CC] Never use simple partial apply for async functions.
Currently, the convention for partially applying an async function is to
form a thick function consisting of (1) a pointer to the partial apply
forwarder and (2) a thick context containing the async context size.

Previously, the partial application of async methods--specifically, for
which the partial application of the non-async analog produced a simple
partial apply (meaning that their context was a single ref-counted
object in the form of self)--would result in a simple partial
application.  Here, that error is addressed by never allowing the
partial application of async methods to produce a simple partial apply.

Eventually, when the complex temporary ABI for partial application of
async functions is unrolled, this change will need to be unrolled as
well.