[CodeCompletion][Sema] Don't guess whether a solution will be applied to the AST based on solutions formed for code completion.
When solving for code completion, we ignore missing arguments after the one
containing the code completion location, don't favor overloads with a number of
params matching the number of arguments in a call and so on. Because of this,
we can't assume multiple solutions being formed when solving for code completion
means there won't be a single best solution formed when solving for regular
type-checking.

Resolves rdar://problem/72362275