[Driver] Add -assert-config as module interface option
Currently the `-assert-config` flag is not serialized into the module interface. This can cause a subtle issue when rebuilding a Swift module from the corresponding .swiftinterface when the module cache has gone stale.

This change adds `-assert-config` as a module interface option so that
it will be serialized into the header of the .swiftinterface file of the
module.

rdar://72452477