Clearly discriminate OperandOwnership::NonUse and TrivialUse.
A NonUse operand does not use the value itself, so it ignores
ownership and does not require liveness. This is for operands that
represent dependence on a type but are not actually passed the value
of that type (e.g. they may refer an open_existential). This could be
used for other dependence-only operands in the future.

A TrivialUse operand has undefined ownership semantics, so it only
handles values with ownership None. It does still require
liveness. The name makes defining ownership for trivial operations
perfectly obvious, increasing readability of OperandOwnership.cpp.