build: improve _InternalSwiftSyntaxParser installation
Use the `TARGETS` form of `install` to avoid having to spell out the
location and file name to install.  Additionally, setup the installation
rules to work better for Windows by listing install locations for the
archive (static library) form (which includes the import library on
Windows), library form (which is the Unix library), and the runtime
component (the .dll on Windows).