SILOptimizer: treat a `destroy_addr` as modification
When performing the r-value elimination pass repeatedly, we may have
inserted a `destroy_addr` of the source value.  In such a case, it is
important to treat the `destroy_addr` as a modification (as you may not
use the value any longer).  This seems to match the behaviour observed
on macOS (and theoretically Windows).  Although this appears to extra
copy, it does not misuse memory.